refactor(tests): extract common tests and trait them
